Refactor: Suffixe in helpers/
