+++ server/helpers/parseInfo.helpers.js
